Move fit_reg function to utils.py
